Avoid bad-slot-type error in CCL
CCL was throwing a bad-slot-type error when %finalize-tiles attempts to set the terrains slot-value in tile.  Some trial and error showed that it really only dislikes the length in the type specifier for terrains in the tiled-tileset-tile class.  Since %finalize-tiles is indeed creating an array of length 4, this looks like it may be a bug in CCL.  This commit uses a reader conditional to work around the issue by leaving the length in the type specifier undefined when working in CCL.